requirements.txt
# 웹 프레임워크
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
starlette>=0.27.0
httpx>=0.24.0

# 비동기 처리
asyncio>=3.4.3
aiohttp>=3.8.1

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
beautifulsoup4>=4.9.3
lxml>=4.9.0

# 벡터 데이터베이스
pinecone-client>=2.2.1
# 대체 벡터 DB 클라이언트 (필요시 주석 해제)
# weaviate-client>=3.15.4
# pymilvus>=2.2.8
# qdrant-client>=1.1.1
# elasticsearch>=8.7.0
# elasticsearch-vector-plugin>=0.1.0

# AI 및 임베딩
openai>=1.3.0

# 유틸리티
python-dotenv==1.0.0
uuid>=1.30
python-multipart>=0.0.6
pyjwt>=2.6.0
tqdm>=4.65.0

# 로깅
loguru>=0.7.0

# 테스트
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0

# 문서화
mkdocs>=1.4.3
mkdocs-material>=9.1.5

# HTTP 요청
requests==2.31.0 